// src/core/types.ts
export type IScope = 'testcase' | 'testrun' | 'testplan' | 'testreport';

// ─── Config ──────────────────────────────────────────────────────────────────
export interface IConfig {
    version?: string;
    github?: IGitHubConfig;
    backend?: IBackendConfig;
    scope?: {
        global?: ITest;
        testcase?: ITest;
        testrun?: ITest;
        testplan?: ITest;
    };
    report_mapping?: Record<string, string>;
}

interface IBackendConfig {
    type: 'local' | 's3' | 'azurerm';
    config: Record<string, string>;
}

export interface IGitHubConfig {
    owner: string;
    repository: string;
    projectId?: number;
    tokenEnv: string;
}

interface IConvention {
    directory?: string;
    filename?: string;
}

export interface ITest {
    identity?: string;
    fields?: IField[];
    convention?: IConvention;
}

export interface IField {
    name: string;
    type: 'keywords' | 'tags';
    values?: string | string[];
    required?: boolean;
    default?: string | string[];
}